requirements.txt
# DMP-SL-Agent: Data Marketplace Semantic Layer Agent
# Python 3.10+ required

# ============================================================================
# Core Dependencies
# ============================================================================

pydantic==2.0.0
pydantic-settings==2.0.0
pyyaml>=6.0
python-dotenv==1.0.0
httpx==0.27.0  # async HTTP client

# ============================================================================
# LangGraph and LangChain
# ============================================================================
# CRITICAL: langchain-core version MUST be exactly 0.3.83
# safechain may install an incompatible version - this is overridden below
# If you have version conflicts, try: langgraph==0.2.50, langchain-core==0.3.28

safechain
langgraph==0.2.50          # V2 agent orchestration
langchain-core==0.3.83     # MUST be this exact version - installed AFTER safechain

# ============================================================================
# MCP SDK (Model Context Protocol)
# ============================================================================
# Official Anthropic MCP SDK for Python

mcp==1.0.0

# SSE client for MCP transport
httpx-sse==0.4.0

# ============================================================================
# Console Output
# ============================================================================

rich==13.0.0  # pretty console output for thinking display
